Basal Disease

Basal Disease

Usage

basal_disease(disease, disease_lookup, encountered_diseases = character())

Arguments

disease

Disease for which to determine basal disease

disease_lookup

Table with two columns – disease and nesting_disease

encountered_diseases

Character vector of diseases already found. Typically this left at the default value of an empty character vector.

Value

The root disease that input disease maps to in disease_lookup.