Convert Harmonized Metadata

Get metadata for a harmonized data source, given metadata for the corresponding tidy data source metadata and initial harmonized data source metadata.

Usage

convert_harmonized_metadata(
  tidy_metadata,
  harmonized_metadata,
  tidy_source,
  harmonized_dataset_id,
  tidy_source_metadata_path
)

Arguments

tidy_metadata

Metadata from read_tracking_tables for a tidy data source.

harmonized_metadata

Initial metadata from read_tracking_tables for a harmonized data source.

tidy_source

IIDDA data source ID for a data source that is being harmonized.

harmonized_dataset_id

ID of dataset being harmonized.

tidy_source_metadata_path

Output of convert_metadata_path.