Data to JSON Files

Create a directory of JSON files from a data frame.

Usage

data_to_json_files(data, json_dir, name_field, use_extension = FALSE)

Arguments

data

Data frame

json_dir

Path to the directory for saving the JSON files.

name_field

Name of the field in the CSV file that contains the names for each JSON file. All values in this field must be unique.

use_extension

If there is a column in the CSV file called `extension`, should it be used to produced json filenames of the form `value-in-name-field.value-in-extension-field.json`?