Create New IIDDA Dataset from Single File

Create New IIDDA Dataset from Single File

Usage

iidda_from_single_file(single_file, new_repo, lifecycle)

Arguments

single_file

path to single data file

new_repo

path to new IIDDA repository

lifecycle

character vector giving the lifecycle state (https://github.com/davidearn/iidda/blob/main/LIFECYCLE.md) Probably 'Unreleased', but it could in principle be 'Static', 'Dynamic', or 'Superseded'.

Value

No return value. Call to produce a new directory structure in a new IIDDA git repository containing a single source data file.