Save Results of a Data Prep Script

Save the resulting objects of a data prep script into an R data file. The names of the resulting objects are given by the names of the result list.

Usage

save_result(result, metadata)

Arguments

result

Named list of data resulting from data prep scripts

metadata

Nested named list describing metadata for the result. It must have a $Product[["Path to tidy data"]] component, which is a GitHub URL describing the ultimate location of the R data file. The GitHub component of the URL will be removed to produce a path that will correspond to the location within a cloned git repository – note that the path is relative to the top-level of the cloned repository.