Compare Two Excel Files

Report on the differences between two xlsx files.

Usage

xlsx_diff(path_one, path_two, ...)

Arguments

path_one

Path to an Excel file.

path_two

Path to an Excel file.

...

Additional arguments to pass to xlsx_cells.

Value

Either `TRUE` if the two files are identical, or a list with the following items. * `all_equal` : Result of applying all.equal to the data frames representing each Excel file. * `in_both_but_different` : Data frame containing cells that are in both Excel files but with different values. * `in_one_only` : Data frame containing cells that are in the first Excel file but not the second. * `in_two_only` : Data frame containing cells that are in the second Excel file but not the first.